+++ drake/solvers/test/moby_lcp_solver_test.cc
@@ -6,7 +6,6 @@
 #include <gtest/gtest.h>
 
 #include "drake/common/eigen_matrix_compare.h"
-#include "drake/util/testUtil.h"
 
 namespace drake {
 namespace solvers {